Fixed some bug of DistributedRadixTree and LocalRadixTree

+++ csrc/redis_meta_channel.h
@@ -78,7 +78,7 @@ public:
   size_t load(std::vector<BlockMeta> &out, size_t max_items);
 
   // Batch update lt for all block metas belonging to node_id
-  bool renew_node_leases(uint32_t node_id, uint32_t new_lt, size_t batch_size = 200);
+  bool renew_node_leases(uint32_t node_id, uint64_t new_lt, size_t batch_size = 200);
 
   // Returns the global node id assigned to this process, or UINT32_MAX if uninitialized.
   uint32_t get_node_id() const;
